requirements.txt
wheel
pandas
selenium
python-dotenv
pysftp
openpyxl
requests
boxsdk[jwt]
gitpython
